requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes -o requirements.txt
-e .
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   openai
azure-common==1.1.28
    # via azure-search-documents
azure-core==1.38.2
    # via
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
azure-identity==1.25.2
    # via ecossistema-capacitacao-grp
azure-search-documents==11.6.0
    # via ecossistema-capacitacao-grp
azure-storage-blob==12.28.0
    # via ecossistema-capacitacao-grp
blinker==1.9.0
    # via flask
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.5
    # via requests
click==8.3.1
    # via flask
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
    #   tqdm
cryptography==46.0.5
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.3
    # via ecossistema-capacitacao-grp
gunicorn==25.1.0
    # via ecossistema-capacitacao-grp
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
isodate==0.7.2
    # via
    #   azure-search-documents
    #   azure-storage-blob
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.13.0
    # via openai
librt==0.8.1 ; platform_python_implementation != 'PyPy'
    # via mypy
lxml==6.0.2
    # via python-docx
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
msal==1.35.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
mypy==1.19.1
mypy-extensions==1.1.0
    # via mypy
openai==2.30.0
    # via ecossistema-capacitacao-grp
openpyxl==3.1.5
    # via ecossistema-capacitacao-grp
packaging==26.0
    # via
    #   gunicorn
    #   pytest
pathspec==1.0.4
    # via mypy
pluggy==1.6.0
    # via pytest
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pyjwt==2.11.0
    # via msal
pytest==9.0.2
python-docx==1.2.0
    # via ecossistema-capacitacao-grp
python-dotenv==1.2.2
    # via ecossistema-capacitacao-grp
requests==2.32.5
    # via
    #   azure-core
    #   msal
ruff==0.15.5
sniffio==1.3.1
    # via openai
tqdm==4.67.3
    # via openai
types-openpyxl==3.1.5.20250919
typing-extensions==4.15.0
    # via
    #   azure-core
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
werkzeug==3.1.6
    # via
    #   ecossistema-capacitacao-grp
    #   flask
